[UPDATE] moving all from first_iteration to OndiraisDeLaMagie
