--- README.md
@@ -1,8 +1,0 @@
-# On Dirait De La Magie !
-
-This branch is your branch. You should work here, and create Pull Requests to `master`
-
-A little Micro-services architecture project: cf [Architecture Graph](https://github.com/rbobillo/OnDiraitDeLaMagie/blob/master/documentation/On%20Dirait%20De%20La%20Magie.pdfhttps://github.com/rbobillo/OnDiraitDeLaMagie/blob/master/documentation/On%20Dirait%20De%20La%20Magie.pdf)
-
-### First Iteration
-Read `first_iteration` [Readme](https://github.com/rbobillo/OnDiraitDeLaMagie/blob/reference/first_iteration/Readme.md) from the `reference` branch
