+++ magic/Readme.md
