+++ magic/dao/Wizard.go
