+++ magic/internal/environment.go
