+++ magic/internal/errors.go
