+++ magic/internal/logging.go
