+++ magic/internal/wizardgenerator.go
@@ -6,7 +6,7 @@ package internal
 
 import (
 	"encoding/json"
-	"github.com/rbobillo/OnDiraitDeLaMagie/first_iteration/magic/dao"
+	"github.com/rbobillo/OnDiraitDeLaMagie/magic/dao"
 	"github.com/satori/go.uuid" // go get github.com/satori/go.uuid
 	"io/ioutil"
 	"math/rand"
