+++ magic/magic.env
